Change repo order for SLE-15-SP1 to fix priorities
Current SLE15SP1 installation fails with:

admin: Problem: vim-data-8.0.1568-3.20.noarch requires \
  vim-data-common = 8.0.1568-3.20, but this requirement \
  cannot be provided

The reason is, that the update repository has a lower priority
than the base repository. But vim-data-common is already installed
from the update repository.

This commit changes the order so packages from the updates repos are
taken instead of packages from the base repos.