Add deepsea grain
Then installing SES6 vagrant-ceph sets everything up but running any
stages fail because the nodes are expected to be identified via a
deepsea grain.

This patch adds a command to admin node to create this grain with:

 salt '*' grains.append deepsea default

I'm not sure when the grain was required, so only added it to the Leap
15.1 and tumbleweed cases. Please advise if more cases are required (or
I missed something).

Signed-off-by: Matthew Oliver <moliver@suse.com>